+++ src/main/java/com/cg/spring/boot/demo/repository/CustomerRepository.java
@@ -7,11 +7,12 @@ import org.springframework.stereotype.Repository;
 import  com.cg.spring.boot.demo.model.Customer;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
 
 
 //	public Customer findByAadhar(Long aadharNumber);
 public List<Customer> findByFirstName(String firstName) ;
 
 public Customer findByMobileNumber(Long mobileNumber);
+
 }
