+++ src/main/java/com/cg/spring/boot/demo/service/CustomerService.java
@@ -10,7 +10,7 @@ import com.cg.spring.boot.demo.model.Customer;
 public interface CustomerService {
 	// check for Duplicate addhar, email, mobile
 	public Customer registerCustomer(Customer customer)throws DuplicateCustomerException;
-	public Customer viewCustomerProfile(Long customerId)throws NoSuchCustomerException;
+	public Customer viewCustomerProfile(int customerId)throws NoSuchCustomerException;
 	//public Customer editCustomerProfile(Integer customerId)throws NoSuchCustomerException;
 	public Customer searchCustomerByCustomerId(Long customerId)throws NoSuchCustomerException;
 	
@@ -20,7 +20,8 @@ public interface CustomerService {
 	//public List<Customer> searchCustomerByName(String customerName)throws NoSuchCustomerException;
 	
 	public List<Customer> searchCustomerByName(String firstName)throws NoSuchCustomerException;
-	Customer editCustomerProfile(Integer customerId) throws NoSuchCustomerException;
+	Customer editCustomerProfile(int customerId) throws NoSuchCustomerException;
+	
 
 	
 //	public Connection newConnectionRequest(Connection connection);
